pipelines/data/requirements.txt
pandas>=1.3.0
numpy>=1.21.0
pyyaml>=5.4.0
transformers>=4.20.0
scikit-learn>=0.24.0
langdetect>=1.0.9
datasets>=2.0.0
